code commit@rahul

+++ src/main/java/com/hrms/controller/EmployeeRequisitionController.java
@@ -3,18 +3,17 @@ package com.hrms.controller;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.HashMap;
+
 import java.util.List;
-import java.util.Map;
+
 
 import javax.servlet.http.HttpSession;
-import javax.validation.Valid;
+
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.InitBinder;
@@ -32,7 +31,6 @@ import com.hrms.helper.Message1;
 import com.hrms.model.CommonUtil;
 import com.hrms.model.MenuModule;
 import com.hrms.model.EmployeeRequisition;
-//import com.hrms.model.EmployeeRequisitionCreationDto;
 import com.hrms.model.EmployeeRequisitionDetail;
 import com.hrms.service.DepartmentService;
 import com.hrms.service.DesignationService;
@@ -44,7 +42,7 @@ import com.hrms.service.PageMappingService;
 @Controller
 public class EmployeeRequisitionController {
 	int  pageNo= 71;
-	String reqPage = "employeeRequisition";  //employeeRequisition.html.html
+	String reqPage = "employeeRequisition";  
 	@Autowired ModuleService moduleService;
 	@Autowired PageMappingService pageMappingService;
 	@Autowired DepartmentService departmentService;
@@ -63,8 +61,7 @@ public class EmployeeRequisitionController {
                 new CustomDateEditor(dateFormatter, true));
         binder.registerCustomEditor(Date.class, "reqTill",
                 new CustomDateEditor(dateFormatter, true));
-//        binder.registerCustomEditor(Date.class, "reqDate",
-//                new CustomDateEditor(dateFormatter, true));
+
        
     }
 	
@@ -72,6 +69,12 @@ public class EmployeeRequisitionController {
 	
 	@GetMapping("employeeRequisition")  
 	public String employeeRequisition(Model model, HttpSession session) {
+		
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
+		
+		
 		session.setAttribute("imgUtil", new ImageUtil());
 		String userCode = (String) session.getAttribute("username");  
 		List<MenuModule> modules = moduleService.getAllModulesList(userCode);
@@ -87,7 +90,7 @@ public class EmployeeRequisitionController {
 	
 		
 		
-		List<EmployeeRequisitionDetail> emp = new ArrayList<EmployeeRequisitionDetail>();
+		List<EmployeeRequisitionDetail> emp = new ArrayList<>();
 		EmployeeRequisition eReq = new EmployeeRequisition();
 		
 		List<EmployeeRequisition> requisition = employeeRequisitionService.getAllEmployeeRequisition();
@@ -95,11 +98,10 @@ public class EmployeeRequisitionController {
 			model.addAttribute("requisition", requisition);
 		}
 		
-		System.out.println("requisition size : "+ requisition.size());
 		
 		
 		eReq.setEmployeRequisitionDetail(emp);
-		System.out.println("lenths =====" + emp.size());
+		
 		model.addAttribute("req", eReq);
 		
 		session.setAttribute("username", session.getAttribute("username"));
@@ -112,8 +114,13 @@ public class EmployeeRequisitionController {
 	public String saveEmployeeRequisition(@ModelAttribute("req") EmployeeRequisition employeeRequisition,
 			HttpSession session,RedirectAttributes redirectAttributes) {
 		
+		
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
+		
 		String insertedBY = (String) session.getAttribute("userlogin");
-
+		
 		try {
 			
 			List<EmployeeRequisitionDetail> re = new ArrayList<>();
@@ -136,46 +143,44 @@ public class EmployeeRequisitionController {
 		}catch(Exception e) {
 			e.printStackTrace();
 			session.setAttribute("message", new Message1("Something went Wrong !!","alert-info"));
-
 			
 		}
 		
 		session.setAttribute("userlogin", insertedBY);
 
-		return "redirect:employeeRequisition"; //+pageMappingService.PageRequestMapping(reqPage,pageNo)
+		return "redirect:employeeRequisition"; 
 	}
 	
 	
 	@GetMapping(value = {"editRequisition/{id}"})
 	public String editRequisition(@PathVariable("id") String reqCode,@ModelAttribute("req") EmployeeRequisition employeeRequisition,
-			Model model) {
-		System.out.println("==========employeeRequisitionCode===========>"+reqCode);
+			Model model,HttpSession session) {
+	
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
 		
 		EmployeeRequisition requisition = employeeRequisitionService.findEmployeeRequisitiondById(reqCode);
-		System.out.println("requisition detail : "+ requisition.getReqCode());
 		
-		System.out.println("requisition detail : "+requisition.getEmployeRequisitionDetail().size());
-		
-		System.out.println("requisition detail : "+requisition.getEmployeRequisitionDetail().get(0).getAgeFrom());
-		
-		//System.out.println("desgnation through requisition : " + requisition.getEmployeRequisitionDetail().get(0).getDesignation().getDesgCode());
 		
 
-		//requisition.getEmployeRequisitionDetail()
+		
 		if(requisition != null) {
 			model.addAttribute("req", requisition);
 		}
-		System.out.println("employee requisition id : "+ reqCode);
+		
 		return "editEmployeeRequisition";
 	}
 	
 	@PostMapping(value = {"updateRequisition"})
 	public String updateRequisition(@ModelAttribute("req") EmployeeRequisition employeeRequisition,
-			Model model) {
+			Model model,HttpSession session) {
 		
-		  System.out.println("=====================>update employee Requisition");
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
 		  
-		ArrayList<EmployeeRequisitionDetail> employeeRequisitionDetail = new ArrayList<EmployeeRequisitionDetail>();
+		ArrayList<EmployeeRequisitionDetail> employeeRequisitionDetail = new ArrayList<>();
 		
 		for(EmployeeRequisitionDetail eDetail : employeeRequisition.getEmployeRequisitionDetail()) {
 			eDetail.setEmployeeRequisition(employeeRequisition);
@@ -189,11 +194,14 @@ public class EmployeeRequisitionController {
 	
 	@GetMapping(value = {"deleteRequisition/{id}"})
 	public String deleteRequisition(@PathVariable("id") String reqCode, Model model, HttpSession session) {
-		System.out.println("=====================>");
+		
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
 		
 		employeeRequisitionService.removeEmployeeRequisition(reqCode);
 		
-		System.out.println("employee requisition id : "+ reqCode);
+		
 		session.setAttribute("username", session.getAttribute("username"));
 		return "redirect:/employeeRequisition";
 	}
@@ -201,15 +209,18 @@ public class EmployeeRequisitionController {
 	
 	@GetMapping(value = {"viewRequisition/{id}"})
 	public String viewRequisition(@PathVariable("id") String reqCode,@ModelAttribute("req") EmployeeRequisition employeeRequisition,
-			Model model) {
-		System.out.println("=====================>");
+			Model model,HttpSession session) {
+	
+		if(session.getAttribute("username")==null) {
+			return "redirect:" + "./";
+		}
 		
 		EmployeeRequisition requisition = employeeRequisitionService.findEmployeeRequisitiondById(reqCode);
-		//requisition.getEmployeRequisitionDetail()
+		
 		if(requisition != null) {
 			model.addAttribute("req", requisition);
 		}
-		System.out.println("employee requisition id : "+ reqCode);
+		
 		return "viewEmployeeRequisition";
 	}
 	
@@ -222,9 +233,9 @@ public class EmployeeRequisitionController {
 	@GetMapping("/designationbydept/{selecteDepartmentValue}")
 	public List<CommonUtil> getDesignationByDepartment(@PathVariable("selecteDepartmentValue") String deptCode) {
 		
-		System.out.println("department value is"+ deptCode);
+		
 		List<Employee> employeeList = employeeService.findByDepartmentCode(deptCode);
-		List<CommonUtil> details = new ArrayList<CommonUtil>();
+		List<CommonUtil> details = new ArrayList<>();
 		CommonUtil empl = new CommonUtil();
 		for(int i =0;i<employeeList.size();i++) {
 			Designation designation = designationService.findDesignationById(employeeList.get(i).getDesignationCode());
@@ -241,7 +252,7 @@ public class EmployeeRequisitionController {
 	@ResponseBody
 	@GetMapping("designationsize")
 	public int getDesignationSize() {
-	System.out.println("ïn designation method");
+	
 		return designationService.getAllDesignations().size();
 	}
